Non-Core services

+++ spec/spec_helper.rb
@@ -5,8 +5,8 @@ require 'vcr'
 ENV['SDK']='openstack'
 
 FLAVOR_ID  = '1'
-IMAGE_ID   = 'd1a3bf61-7346-482e-91aa-b038ad784042'
-PROJECT_ID = '92ac5831e22341bc9cc88f3942ba4d68'
+IMAGE_ID   = ''
+PROJECT_ID = ''
 
 RSpec.configure do |config|
   config.expect_with :rspec do |expectations|
@@ -31,16 +31,6 @@ RSpec.configure do |config|
     c.filter_sensitive_data('<OS_TENANT_NAME>') { ENV['OS_TENANT_NAME'] }
   end
 
-  # We must keep this at :each recurrence; otherwise we swamp the testing server limits on large enough test runs.
-  config.before(:each) do
-    if VCR.current_cassette.recording?
-      OpenStackRubySDK::Cinder::Snapshot.all.each{ |v| v.destroy rescue next }
-      OpenStackRubySDK::Cinder::Backup.all.each{ |v| v.destroy rescue next }
-      OpenStackRubySDK::Cinder::Volume.all.each{ |v| v.destroy rescue next }
-      OpenStackRubySDK::Nova::Server.all.each{ |s| s.destroy rescue next }
-    end
-  end
-
   def vcr_on
     VCR.turn_on!
     WebMock.disable_net_connect!
@@ -52,9 +42,7 @@ RSpec.configure do |config|
   end
 
   def do_it_live
-    unless recording?
-      File.delete(VCR.current_cassette.file)
-    end
+    File.delete(VCR.current_cassette.file) unless recording?
   end
 
   def recording?
@@ -69,16 +57,14 @@ RSpec.configure do |config|
     server = OpenStackRubySDK::Nova::Server.create({
       name: Time.now.usec.to_s,
       flavorRef: FLAVOR_ID,
-      imageRef: IMAGE_ID
-    })
+      imageRef: IMAGE_ID })
     Peace::Helpers.wait_for(server)
   end
 
   def fresh_volume
     volume = OpenStackRubySDK::Cinder::Volume.create({
       size: '1',
-      name: Time.now.usec.to_s
-    })
+      name: Time.now.usec.to_s })
     Peace::Helpers.wait_for(volume, "available")
   end
 
@@ -89,4 +75,5 @@ RSpec.configure do |config|
     Peace::Helpers.wait_for(v, "in-use")
     Peace::Helpers.wait_for(s)
   end
+
 end
